Add world generation and collision
World generation via my Zig Perlin library
<github.com/mgord9518/perlin-zig> and collision detection is implemented
by checking if the middle of the screen (where the player stands) overlaps with a tile, then checking if the pixel to one of the sides also overlaps. If eg: the left pixel doesn't overlap as well, the players X value will be subtracted. This is currently really buggy due to the variable speed the player can move at and I'll fix it by adding a speed variable to the player and simply reversing it if hitting a wall