Greatly improve collision
Due to now calculating the collision based on rectangles instead of a
single point, the player now properly slides across walls. There is
still a bug which catches the player on corners when sliding up, but I
intend to fix this next.

Once it gets working I'll refactor the collision code.

As another fix, the menu drawing code has been moved into its own
function and there is now a variable to enable pixel snapping (previous
behavior) and it is disabled by default to make the game feel smoother