Fix text display, remove zig-toml
zig-toml is quite out of date and I don't have the time to update it to
the latest Zig, so I'll probably use a different format for config
files. Maybe JSON.

Text display suddenly broke, causing lines to be misaligned, not sure if
this is a problem with raylib or my code, but it only misbehaves using
scaled text. I just made a simple helper function to properly space with
scaling in mind, and it was needed anyway so that I could easily print
text of the same style.

I also did a bit of cleaning up, but it has a far way to go. I probably
won't add any real features until the engine is somewhat solid